moved test utils for driver into own file
